[refactor][QuarantineStatus] Separate time and day in object fields
**Why this change was necessary**
Having both day count and time of day as entries in the same tuple in
one field is poor design.

**What this change does**
Creates a new field just for the day count and changes the time of day
field to only contain time of day rather than a tuple.

Changes any accesses to those fields in output strings.

**Any side-effects?**
There is a multiple assignment when next() is called on the time of
day generator.

**Additional context/notes/links**

Resolves #25 - Separate time of day from day count in object